initial docker support
If goodplay_image variable with docker image name is specified on an inventory host the appropriate docker image is pulled (if not already available) and containers are started prior to testing. A dynamic inventory is created which configures the connection to the docker containers. Thus goodplay testing can be effectively executed against a whole set of docker containers simulating a real infrastructure. After testing containers are automatically destroyed.