add missing ``ansible_user`` inventory variable in tests
Since Ansible 2.0 there is no default remote user defined anymore (prior to that it was the current user). Up to now this didn't occur as a problem as the ``ansible_user`` wasn't used with the docker connection plugin. But this behavior has been changed in recent ansible commit ``14dfad7``.

The goodplay documentation has been also updated to reflect the change.